requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0
feedparser>=6.0.10

# YouTube processing
youtube-transcript-api>=0.6.1
pytube>=15.0.0

# AI and summarization
openai>=1.3.0
anthropic>=0.7.0

# Scheduling and async
schedule>=1.2.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Data storage and processing (Optional - only needed for local storage)
# sqlalchemy>=2.0.0
# alembic>=1.13.0
pandas>=2.1.0
pydantic>=2.5.0

# Web scraping utilities
newspaper3k>=0.2.8
readability-lxml>=0.8.1

# CLI and configuration
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
rich>=13.7.0

# Notion integration
notion-client>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
